feat: add form field hint and make display name optional
This commit introduces a new `hint` property for form fields, which
is rendered as a placeholder in the input element.

The `displayName` property for `FieldConfig` has been made optional.
The ultimate form component now intelligently defaults `displayName`
to the `name` of the field if `displayName` is not explicitly provided
in the field configuration.

An example of the new `hint` property and removal of redundant
`displayName` definitions is provided in `app.ts`.